fix: add proper errors for invalid user inputs
Currently, incorrect or invalid URL query parameters were ignored,
giving unexpected responses and lists of reviews.  For example, when an
ExclusiveStartKey required a createdAt property, but it isn't included,
the results returned will be as though an ExclusiveStartKey was not
given.  This may lead users to think that the list of reviews returned
is the next page of results, when it isn't.